Restructuring code
The project now include 3 projects and 1 shared project. The BouncingBall project is a dependency which load settings for the two others. The shared Project contains all classes uses in both last projects. Finally the BrokerApplication launch only the broker part of the project and the ClientApplication only the Client part.